:recycle: refactor: change way to import css in component
