+++ src/components/Badge/Badge.ts
@@ -1,6 +1,6 @@
-import { LitElement, html, unsafeCSS } from "lit";
+import { LitElement, html, unsafeCSS, css } from "lit";
 import { customElement, property } from "lit/decorators.js";
-import BadgeStyle from "./Badge.css" assert { type: "css" };
+import BadgeStyle from "./Badge.css?inline";
 
 export type CandyBadgeProps = {
   label: string;
@@ -8,7 +8,9 @@ export type CandyBadgeProps = {
 
 @customElement("candy-badge")
 export class CandyBadge extends LitElement {
-  static styles = unsafeCSS(BadgeStyle);
+  static styles = css`
+    ${unsafeCSS(BadgeStyle)}
+  `;
 
   @property({ type: String })
   label = "Badge";
