+++ src/components/Divider/Divider.ts
@@ -1,10 +1,12 @@
-import { LitElement, html, unsafeCSS } from "lit";
+import { LitElement, html, unsafeCSS, css } from "lit";
 import { customElement } from "lit/decorators.js";
-import DividerStyle from "./Divider.css" assert { type: "css" };
+import DividerStyle from "./Divider.css?inline";
 
 @customElement("candy-divider")
 export class CandyDivider extends LitElement {
-  static styles = unsafeCSS(DividerStyle);
+  static styles = css`
+    ${unsafeCSS(DividerStyle)}
+  `;
 
   render() {
     return html`<div part="divider" class="divider-container"><div></div></div> `;
