+++ src/components/Link/Link.ts
@@ -1,6 +1,6 @@
-import { LitElement, html, unsafeCSS } from "lit";
+import { LitElement, html, unsafeCSS, css } from "lit";
 import { customElement, property } from "lit/decorators.js";
-import LinkStyle from "./Link.css" assert { type: "css" };
+import LinkStyle from "./Link.css?inline";
 
 export type CandyLinkProps = {
   label: string;
@@ -10,7 +10,9 @@ export type CandyLinkProps = {
 
 @customElement("candy-link")
 export class CandyLink extends LitElement {
-  static styles = unsafeCSS(LinkStyle);
+  static styles = css`
+    ${unsafeCSS(LinkStyle)}
+  `;
 
   @property({ type: String })
   label = "click me";
