+++ src/components/Switch/Switch.ts
@@ -1,6 +1,6 @@
-import { LitElement, html, unsafeCSS } from "lit";
+import { LitElement, html, unsafeCSS, css } from "lit";
 import { customElement, property } from "lit/decorators.js";
-import SwitchStyle from "./Switch.css" assert { type: "css" };
+import SwitchStyle from "./Switch.css?inline";
 
 export type CandySwitchProps = {
   activated: boolean;
@@ -8,7 +8,9 @@ export type CandySwitchProps = {
 
 @customElement("candy-switch")
 export class CandySwitch extends LitElement {
-  static styles = unsafeCSS(SwitchStyle);
+  static styles = css`
+    ${unsafeCSS(SwitchStyle)}
+  `;
 
   @property({ type: Boolean })
   activated = false;
